Great, got the exception correctly extending.
It seems that both TypeScript and Babel have changed the way their extenders work. This is fine
for when extending an ordinary class, but the Error class has a weird setup. To get around this,
I've used the util.inherits to extend the Error class, rather than the more traditional 'extends Error'